requirements.txt
transformers
flask
numpy
torch
torchvision
faiss-cpu
openai
streamlit
Pillow
pandas
clip-anytorch
opencv-python-headless
matplotlib
scikit-learn
flask
nltk
sentence_transformers
